third_party/WebKit/LayoutTests/fast/media/media-query-list-missing-arguments.html

<!doctype html>
<script src="../../resources/js-test.js"></script>
<script>
var mediaList = matchMedia("(max-width: 100px)");
shouldBeType(mediaList, "MediaQueryList");
shouldThrow("mediaList.addListener()");
shouldThrow("mediaList.removeListener()");
</script>
