third_party/blink/web_tests/platform/generic/fast/media/media-query-list-missing-arguments-expected.txt

PASS [object MediaQueryList] is an instance of MediaQueryList
PASS mediaList.addListener() threw exception TypeError: Failed to execute 'addListener' on 'MediaQueryList': 1 argument required, but only 0 present..
PASS mediaList.removeListener() threw exception TypeError: Failed to execute 'removeListener' on 'MediaQueryList': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
